+++ example/game/src/main/java/no/nordicsemi/android/ble/example/game/client/repository/ClientConnection.kt
@@ -12,7 +12,7 @@ import no.nordicsemi.android.ble.BleManager
 import no.nordicsemi.android.ble.example.game.client.data.Request
 import no.nordicsemi.android.ble.example.game.proto.OpCodeProto
 import no.nordicsemi.android.ble.example.game.proto.RequestProto
-import no.nordicsemi.android.ble.example.game.quiz.repository.Question
+import no.nordicsemi.android.ble.example.game.server.repository.Question
 import no.nordicsemi.android.ble.example.game.server.data.Players
 import no.nordicsemi.android.ble.example.game.server.data.Results
 import no.nordicsemi.android.ble.example.game.spec.DeviceSpecifications
