+++ example/game/src/main/java/no/nordicsemi/android/ble/example/game/server/repository/QuestionRepository.kt
@@ -1,7 +1,7 @@
-package no.nordicsemi.android.ble.example.game.quiz.repository
+package no.nordicsemi.android.ble.example.game.server.repository
 
-import no.nordicsemi.android.ble.example.game.quiz.api.QuestionsService
-import no.nordicsemi.android.ble.example.game.quiz.model.QuestionMapper
+import no.nordicsemi.android.ble.example.game.server.api.QuestionsService
+import no.nordicsemi.android.ble.example.game.server.model.QuestionMapper
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -9,7 +9,7 @@ import javax.inject.Singleton
 class QuestionRepository @Inject constructor(
     private val service: QuestionsService
 ) {
-    suspend fun getQuestions(category: Int?): Questions{
+    suspend fun getQuestions(category: Int?): Questions {
         val questionsRemote = service.getQuestions(10, category)
         return QuestionMapper.mapRemoteToLocal(questionsRemote)
     }
