+++ example/game/src/main/java/no/nordicsemi/android/ble/example/game/server/view/PlayersNameDialog.kt
@@ -1,4 +1,4 @@
-package no.nordicsemi.android.ble.example.game.quiz.view
+package no.nordicsemi.android.ble.example.game.server.view
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
