+++ example/game/src/main/java/no/nordicsemi/android/ble/example/game/server/view/ResultView.kt
@@ -1,4 +1,4 @@
-package no.nordicsemi.android.ble.example.game.quiz.view
+package no.nordicsemi.android.ble.example.game.server.view
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.*
